Refactor: Split formatter.rs into smaller modules. Closes #20
