+++ src/compiler/frontend/formatter/ty.rs
@@ -1,0 +1,42 @@
+use crate::compiler::ast::*;
+use crate::compiler::frontend::formatter::Formatter;
+
+pub(crate) fn format_type_expr(f: &mut Formatter, ty: &TypeExpr) {
+    match ty {
+        TypeExpr::Name(name, _) => f.result.push_str(name),
+        TypeExpr::Union(tys, _) => {
+            for (i, t) in tys.iter().enumerate() {
+                if i > 0 {
+                    f.result.push_str(" | ");
+                }
+                format_type_expr(f, t);
+            }
+        }
+        TypeExpr::Generic(name, args, _) => {
+            f.result.push_str(name);
+            f.result.push('<');
+            for (i, arg) in args.iter().enumerate() {
+                if i > 0 {
+                    f.result.push_str(", ");
+                }
+                format_type_expr(f, arg);
+            }
+            f.result.push('>');
+        }
+        TypeExpr::Array(item, _) => {
+            format_type_expr(f, item);
+            f.result.push_str("[]");
+        }
+        TypeExpr::Function(params, ret, _) => {
+            f.result.push_str("function(");
+            for (i, p) in params.iter().enumerate() {
+                if i > 0 {
+                    f.result.push_str(", ");
+                }
+                format_type_expr(f, p);
+            }
+            f.result.push_str("): ");
+            format_type_expr(f, ret);
+        }
+    }
+}
